backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["uday kiran <udaykiranchintu16@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dotenv = "^1.0.1"
bcrypt = "^4.2.0"
passlib = "^1.7.4"
sqlalchemy = "^2.0.35"
pydantic = {extras = ["email"], version = "^2.9.2"}
regex = "^2024.9.11"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.5.2"
python-multipart = "^0.0.12"
pymongo = "^4.10.1"
dnspython = "^2.7.0"
google-cloud-storage = "^2.18.2"
openai = "^1.51.2"
mutagen = "^1.47.0"
python-pptx = "^1.0.2"
docx = "^0.2.4"
pypdf2 = "^3.0.1"
pandas = "^2.2.3"
python-docx = "^1.1.2"
llama-index = "^0.11.17"
llama-index-indices-managed-llama-cloud = "^0.3.1"
docx2txt = "^0.8"
xhtml2pdf = "^0.2.14"
markdown = "^3.7"
e2b_code_interpreter = "0.0.11b38"
nest-asyncio = "^1.6.0"
jupyter = "^1.1.1"
aiostream = "^0.6.4"
pymilvus = "^2.4.9"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"